IR: Split parser from IRLoader, add IR::Parse()
